# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GigFolio API — Developer 3 Scope
  description: |
    API specifications for Developer 3 deliverables in the GigFolio platform:
    - **Reputation Engine:** Breakdown and composite scoring
    - **Consent Management & Share Links:** Time-boxed, scope-limited access delegation
    - **QR Code Generation:** Authoritative backend generation for portable profile sharing
    - **Enterprise Verification:** Scoped data projection and SD-JWT presentation
    - **Application Audit Events:** Tamper-resistant, hashed-identifier access logs
  version: 1.0.0
  contact:
    name: GigFolio Engineering Team
    email: dev@gigfolio.io
servers:
  - url: https://api.gigfolio.io
    description: Production API
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/v1/organizations/verify-worker:
    post:
      tags:
        - Enterprise Verification
      summary: Enterprise worker verification
      description: |
        Verifies a worker profile using a share token. Returns ONLY consented
        data fields (projected at query time) accompanied by a key-bound
        SD-JWT presentation.
      operationId: verifyWorker
      security:
        - OrgApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyWorkerRequest'
            example:
              share_token: raw-token-from-qr-or-link
              access_passcode: optional
              verifier_nonce: cryptographic-challenge-string
      responses:
        '200':
          description: Worker successfully verified and consented claims projected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyWorkerResponse'
              example:
                verification_event_id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                worker:
                  legal_name: Jane Worker
                verification_tier: PLATFORM_VERIFIED
                verified_metrics:
                  composite_score: 782
                  confidence_tier: HIGH_CONFIDENCE
                sd_jwt_presentation: eyJhbGciOiJFUzI1NiJ9.eyJzdWIiOiJ3b3JrZXIifQ.SIGNATURE~DISCLOSURE~
        '401':
          description: Invalid or missing organization API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Wrong passcode, or consent revoked/does not authorize this access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Token expired or usage limit already exhausted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Worker JWT containing worker ID and granted scopes.
    OrgApiKeyAuth:
      type: apiKey
      in: header
      name: X-Organization-Key
      description: Institutional Verifier API Key.
  schemas:
...
    VerifyWorkerRequest:
      type: object
      required:
        - share_token
        - verifier_nonce
      properties:
        share_token:
          type: string
          example: raw-token-from-qr-or-link
        access_passcode:
          type: string
          nullable: true
          example: optional
        verifier_nonce:
          type: string
          example: cryptographic-challenge-string
        worker_public_id:
          type: string
          format: uuid
          nullable: true
          description: >-
            Optional public ID binding. When provided, the backend validates that
            the share_token belongs to this exact worker, preventing cross-worker
            token confusion attacks.
          example: 7f3a2b1c-4e5d-6f7a-8b9c-0d1e2f3a4b5c
    VerifyWorkerResponse:
      type: object
      required:
        - verification_event_id
        - worker
        - verification_tier
        - verified_metrics
        - sd_jwt_presentation
      properties:
        verification_event_id:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        worker:
          type: object
          description: Only consented identity fields
          additionalProperties: true
        verification_tier:
          type: string
          enum: [SELF_REPORTED, PLATFORM_VERIFIED, GOVERNMENT_VERIFIED]
          example: PLATFORM_VERIFIED
        verified_metrics:
          type: object
          description: Only consented performance metrics
          additionalProperties: true
        sd_jwt_presentation:
          type: string
          example: eyJhbGciOiJFUzI1NiJ9.eyJzdWIiOiJ3b3JrZXIifQ.SIGNATURE~DISCLOSURE~
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
          nullable: true
...
  responses:
...
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'